@@ -29,7 +29,7 @@ public class WetoolSupporter implements WePluginSupporter {
     }
 
     private void openTab() {
-        Node node = FxUtils.loadFxml("/ease/devtool/redis/Main.fxml");
+        Node node = FxUtils.loadFxml(WetoolSupporter.class, "/ease/devtool/redis/Main.fxml", true);
         FxUtils.openTab(node, CommonConsts.APP_ID, CommonConsts.APP_NAME);
     }
 }
