@@ -1,11 +1,12 @@
 package org.code4everything.wetool.plugin.devtool.ssh.ssh;
 
-import cn.hutool.extra.ssh.Sftp;
 import lombok.experimental.UtilityClass;
 import org.code4everything.boot.base.ReferenceUtils;
 import org.code4everything.wetool.plugin.devtool.ssh.config.ServerConfiguration;
+import org.code4everything.wetool.plugin.devtool.ssh.hutool.Sftp;
 
 import java.lang.ref.WeakReference;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -23,6 +24,10 @@ public class SftpUtils {
 
     public static ServerConfiguration getConf(String alias) {
         return CONFIGURATION_MAP.get(alias);
+    }
+
+    public static Collection<ServerConfiguration> listConf() {
+        return CONFIGURATION_MAP.values();
     }
 
     public static void putConf(ServerConfiguration server) {
