@@ -6,7 +6,6 @@ import cn.hutool.core.thread.ThreadUtil;
 import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.StrUtil;
-import cn.hutool.extra.ftp.Ftp;
 import javafx.scene.control.ComboBox;
 import javafx.util.Pair;
 import lombok.experimental.UtilityClass;
@@ -18,6 +17,7 @@ import org.code4everything.wetool.plugin.ftp.client.config.FtpConfig;
 import org.code4everything.wetool.plugin.ftp.client.config.FtpInfo;
 import org.code4everything.wetool.plugin.ftp.client.constant.FtpConsts;
 import org.code4everything.wetool.plugin.ftp.client.controller.FtpController;
+import org.code4everything.wetool.plugin.ftp.client.hutool.Ftp;
 import org.code4everything.wetool.plugin.ftp.client.model.FtpDownload;
 import org.code4everything.wetool.plugin.ftp.client.model.FtpUpload;
 import org.code4everything.wetool.plugin.ftp.client.model.LastUsedInfo;
