@@ -38,7 +38,7 @@ public class WetoolSupporter implements WePluginSupporter {
     public MenuItem registerBarMenu() {
         return FxUtils.createMenuItem("插件示例", (EventHandler<ActionEvent>) event -> {
             // 注意保证fxml文件的url路径唯一性
-            Node node = FxUtils.loadFxml("/ease/sample/Sample.fxml");
+            Node node = FxUtils.loadFxml(WetoolSupporter.class, "/ease/sample/Sample.fxml", true);
             FxDialogs.showInformation(SampleController.TAB_NAME, "welcome to wetool plugin");
             FxUtils.openTab(node, SampleController.TAB_ID, SampleController.TAB_NAME);
         });
