@@ -5,7 +5,7 @@ import org.code4everything.wetool.plugin.support.config.WePluginInfo;
 class WetoolTesterTest {
 
     public static void main(String[] args) {
-        WePluginInfo info = new WePluginInfo("ease", "test", "1.0.2", "1.0.2", "");
+        WePluginInfo info = new WePluginInfo("ease", "test", "1.1.0", "1.1.0", "");
         info.setSupportedClass("org.code4everything.wetool.plugin.test.WetoolSupporter");
         WetoolTester.runTest(info, args);
     }
